DL-compliant version from Jim Balhoff and Hilmar Lapp of NESCent.  This version eliminates the use of cardinality restrictions in non-simple properties (ones that are transitive or that are based on a property chain).
On Mar 29, 2011, at 10:16 PM, Jim Balhoff wrote:

Hi all,

Hilmar and I have been looking at what would be required to make
CDAO compliant with OWL 2 DL. It would be nice to be able to use
CDAO with the standard OWL reasoners, but the current file I'm using
(which I think is CDAO v1) includes some axioms which are not
allowed in OWL DL. These all involve the use of non-simple
properties in cardinality restrictions. Here is the definition of a
simple property for OWL DL:

http://www.w3.org/TR/2009/REC-owl2-syntax-20091027/#def_simple

We have made some changes which we think retain all the intended
semantics but express these concepts in a DL-compliant manner. I've
created a table which compares the old and new class definitions:

https://www.nescent.org/wg_evoinfo/CDAO_changes_for_DL-compliance

I also uploaded the edited CDAO file:

https://www.nescent.org/wg/evoinfo/images/5/56/Cdao-dl.owl.zip

Without making some changes to be DL-compliant, importing CDAO into
any other ontologies or datasets makes it impossible to use standard
reasoners. If it is deemed useful to continue adding statements that
go beyond DL, perhaps a DL version of CDAO could be provided, and
then another file which imports that one and contains additional
axioms could be provided.

The file I uploaded works with the HermiT reasoner. However FaCT++
has an issue with the anySimpleType datatype. I think this may be a
FaCT++ deficiency so I haven't yet tried to make any changes
addressing that.

Thanks,
Jim
with