Fixes spelling of JukesCantor label
Fixes #2. Note that this purposely leaves other files intact that
contain the same mis-spelling, specifically `cdao-deprecated-mappings.owl`.
This is because there the mis-spelling is in the term IRIs for v1.0
CDAO term IRIs. (One example why using term labels as part of term IRIs
is can be a bad idea.)